float

+++ train_AE.py
@@ -113,7 +113,7 @@ def main(args):
             encoder.zero_grad()
             decoder.zero_grad()
             # convert to pytorch tensors and Varialbes
-            bobs = torch.from_numpy(obstacles[i:i+args.batch_size])
+            bobs = torch.from_numpy(obstacles[i:i+args.batch_size]).type(torch.FloatTensor)
             bobs = to_var(bobs).view(len(bobs), -1, 3).permute(0,2,1)
             print('data:')
             print(bobs)
